requirements.txt
flask==2.0.1
werkzeug==2.0.1
tushare==1.2.89
pandas==1.3.3
plotly==5.3.1
python-dotenv==0.19.0
numpy==1.21.2 